==> notebooks/4a_generation_model_genetic_training.ipynb <==
import os
os.chdir('..')

from datetime import datetime
from src.models.train_animation_predictor import main

## 1. Set algorithm hyperparameters

# Whether or not to drop paths that shouldn't be animated according to path selector from training
drop = True

# Whether or not to apply path selector and scaling to input data before training
m1 = True

# Number of agents in the genetic algorithm
n_agents = 100

# Number of top agents to be considered in subsequent generations
n_top_parents = 20

# Number of generations to train for
generations = 35

# Paths to training and test datasets
train_path = 'data/path_selector/path_selector_train.csv'
test_path = 'data/path_selector/path_selector_test.csv'

# Timestamp which is used as an ID for saving best models
timestamp = datetime.now().strftime('%Y%m%d_%H%M')

## 2. Start training

best_model = main(train_path=train_path, test_path=test_path, num_agents=n_agents, top_parent_limit=n_top_parents, 
                  generations=generations, drop=drop, timestamp=timestamp, model1=m1)
